Corrgi as validações e o padrão no módulo tarefas, lojas e usuário.

+++ api/modules/colecoes/ColecaoLojaEmBDR.php
@@ -15,38 +15,47 @@ class ColecaoLojaEmBDR implements ColecaoLoja
 	function __construct(){}
 
 	function adicionar(&$obj) {
-		try {	
-			$id = Db::table(self::TABELA)->insertGetId(['razaoSocial' => $obj->getRazaoSocial(), 'nomeFantasia' => $obj->getNomeFantasia()]);
-			
-			$obj->setId($id);
+		if($this->validarLoja($obj)){
+			try {	
+				$id = Db::table(self::TABELA)->insertGetId(['razaoSocial' => $obj->getRazaoSocial(), 'nomeFantasia' => $obj->getNomeFantasia()]);
+				
+				$obj->setId($id);
 
-			return $obj;
+				return $obj;
+			}
+			catch (\Exception $e)
+			{
+				throw new ColecaoException($e->getMessage(), $e->getCode(), $e);
+			}
 		}
-		catch (\Exception $e)
-		{
-			throw new ColecaoException($e->getMessage(), $e->getCode(), $e);
-		}
+		
 	}
 
 	function remover($id) {
-		try {	
-			return DB::table(self::TABELA)->where('id', $id)->delete();
+		if($this->validarDeleteLoja($id)){
+
+			try {	
+				return DB::table(self::TABELA)->where('id', $id)->delete();
+			}
+			catch (\Exception $e)
+			{
+				throw new ColecaoException($e->getMessage(), $e->getCode(), $e);
+			}
 		}
-		catch (\Exception $e)
-		{
-			throw new ColecaoException($e->getMessage(), $e->getCode(), $e);
-		}
+		
 	}
 
 	function atualizar(&$obj) {
-		try {	
-			DB::table(self::TABELA)->where('id', $obj->getId())->update(['razaoSocial' => $obj->getRazaoSocial(), 'nomeFantasia' => $obj->getNomeFantasia()]);
+		if($this->validarLoja($obj)){
+			try {	
+				DB::table(self::TABELA)->where('id', $obj->getId())->update(['razaoSocial' => $obj->getRazaoSocial(), 'nomeFantasia' => $obj->getNomeFantasia()]);
 
-			return $obj;
-		}
-		catch (\Exception $e)
-		{
-			throw new ColecaoException($e->getMessage(), $e->getCode(), $e);
+				return $obj;
+			}
+			catch (\Exception $e)
+			{
+				throw new ColecaoException($e->getMessage(), $e->getCode(), $e);
+			}
 		}
 		
 	}
@@ -106,6 +115,39 @@ class ColecaoLojaEmBDR implements ColecaoLoja
 
     function contagem() {
 		return Db::table(self::TABELA)->count();
+	}
+
+	private function validarLoja(&$obj) {
+		if(!is_string($obj->getRazaoSocial())) {
+			throw new ColecaoException('Valor inválido para razão social.');
+		}
+
+		$quantidade = DB::table(self::TABELA)->where('razaoSocial', $obj->getRazaoSocial())->where('nomeFantasia', $obj->getNomeFantasia())->where('id', '<>', $obj->getId())->count();
+
+		if($quantidade > 0){
+			throw new ColecaoException('Já exite uma loja cadastrada com esses dados.');
+		}
+
+		if(strlen($obj->getRazaoSocial()) <= Loja::TAM_TEXT_MIM && strlen($obj->getRazaoSocial()) > Loja::TAM_TEXT_MAX) throw new ColecaoException('O título deve conter no mínimo '. Loja::TAM_TEXT_MIM . ' e no máximo '. Loja::TAM_TEXT_MAX . '.');
+		if(strlen($obj->getNomeFantasia()) <= Loja::TAM_TEXT_MIM && strlen($obj->getNomeFantasia()) > Loja::TAM_TEXT_MAX) throw new ColecaoException('O nome a fantasia deve conter no mínimo '. Loja::TAM_TEXT_MIM . ' e no máximo '. Loja::TAM_TEXT_MAX . '.');
+
+		return true;
+	}
+
+	private function validarDeleteLoja($id){
+		$qtdReacionamento = DB::table(ColecaoTarefaEmBDR::TABELA)->where('loja_id', $id)->count();
+
+		if($qtdReacionamento > 0){
+			throw new ColecaoException('Essa loja possue tarefas relacionados a ela! Exclua todos as tarefas cadastros e tente novamente.');
+		}
+
+		$qtdReacionamento = DB::table(ColecaoUsuarioEmBDR::TABELA)->where('loja_id', $id)->count();
+
+		if($qtdReacionamento > 0){
+			throw new ColecaoException('Essa loja possue usuários relacionados a ela! Exclua todos os usuários cadastros e tente novamente.');
+		}
+
+		return true;
 	}
 }
